requirements.txt
anthropic==0.42.0
autogen-agentchat==0.2.40
pandas==2.2.3
python-dotenv==1.0.1
openai==1.58.1